+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 * Enums now use strings instead of ints to match thriftrw
 * Enums use Object.freeze to avoid mutations.
+* typedefs of enums are now also values.